src/modules/85-cv/pages/health-source/connectors/CommonHealthSource/components/CustomMetricForm/components/LogsTable/components/JsonSelectorWithDrawer.utils.ts
import { getMultiTypeFromValue } from '@harness/uicore'
import type { GetMultiTypeRecordInitialValueParams, LogFieldsMultiTypeState } from '../../../CustomMetricForm.types'

export const getMultiTypeRecordInitialValue = ({
  filteredFieldsMapping,
  isTemplate,
  formValues
}: GetMultiTypeRecordInitialValueParams): LogFieldsMultiTypeState | null => {
  if (!filteredFieldsMapping || !isTemplate || !formValues) {
    return null
  }

  const logFieldsMultiType: LogFieldsMultiTypeState = {} as LogFieldsMultiTypeState

  filteredFieldsMapping.forEach(field => {
    logFieldsMultiType[field.identifier] = getMultiTypeFromValue(formValues[field.identifier])
  })

  return logFieldsMultiType
}
